requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
pymysql==1.1.0
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
